Remove burn addresses
Tokens are burned by setting the `vout` of an `OutputEntry` to `None` instead of `Some(u32)`.

`OutputEntry` itself has been moved out of proof.rs into a dedicated file.
